+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## ???
+
+* Builtin Prosody: use mod_muc_http_defaults to set rooms properties and prevent unauthorized room creation.
+
 ## v2.0.3
 
 * Fix Peertube server crash when prosody is not installed