+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Features
 
 * Builtin Prosody: use mod_muc_http_defaults to set rooms properties and prevent unauthorized room creation.
+* Builtin Prosody: use mod_auth_http to auto login peertube users
 
 ### Fixes